requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
altair
amqp
annotated-types
anyio
asgiref
asttokens
asyncpg
attrs
autobahn
Automat
backoff
bcrypt
beautifulsoup4
billiard
black
blinker
build
cachetools
celery
certifi
cffi
channels
chardet
charset-normalizer
chroma-hnswlib
chromadb
click
click-didyoumean
click-plugins
click-repl
coloredlogs
constantly
cron-descriptor
cryptography
daphne
dataclasses-json
decorator
deepdiff
Deprecated
deprecation
diskcache
distro
dj-rest-auth
Django
django-allauth
django-celery-beat
django-cors-headers
django-cron
django-crontab
django-filter
django-timezone-field
djangorestframework
djangorestframework-api-key
djangorestframework-simplejwt
docker
docutils
drf-yasg
duckduckgo_search
ecdsa
emoji
executing
faiss-cpu.post1
fastapi
ffmpeg
ffmpeg-python
filelock
filetype
FLAML
flatbuffers
frozenlist
fsspec
future
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
gotrue
greenlet
grpcio
grpcio-status
grpcio-tools
h11
h2
hpack
httpcore
httplib2
httptools
httpx
huggingface-hub
humanfriendly
hyperframe
hyperlink
idna
imageio
imageio-ffmpeg
importlib_metadata
importlib_resources
incremental
inflection
ipython
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
Jinja2
jiter
joblib
jose
jq
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
jwt
keyring
kombu
kubernetes
lancedb
langchain
langchain-chroma
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-openai
langchain-postgres
langchain-qdrant
langchain-text-splitters
langchainhub
langdetect
langgraph
langgraph-checkpoint
langgraph-checkpoint-postgres
langgraph-checkpoint-sqlite
langsmith
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib-inline
mdurl
mmh3
monotonic
more-itertools
moviepy
mpmath
multidict
mypy-extensions
narwhals
nest-asyncio
nh3
nltk
numpy
oauthlib
onnxruntime
openai
openapi-python-client
opencv-python
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentationb0
opentelemetry-instrumentation-asgib0
opentelemetry-instrumentation-fastapib0
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventionsb0
opentelemetry-util-httpb0
ordered-set
orjson
outcome.post0
overrides
packaging
pandas
parso
pathspec
pexpect
pgvector
pillow
pkginfo
platformdirs
playwright
portalocker
postgrest
posthog
primp
proglog
prompt_toolkit
proto-plus
protobuf
psutil
psycopg
psycopg-binary
psycopg-c
psycopg-pool
psycopg2-binary
ptyprocess
pure_eval
py
pyarrow
pyasn1
pyasn1_modules
pyautogen
pyclipper.post5
pycparser
pydantic
pydantic_core
pydeck
pydub
pyee
Pygments
PyJWT
pylance
pyOpenSSL
pyparsing
pypdf
PyPDF2
PyPika
pyproject_hooks
PySocks
python-crontab
python-dateutil.post0
python-decouple
python-docx
python-dotenv
python-iso639
python-jose
python-magic
python-multipart
python-pptx
pytz
PyYAML
qdrant-client
rapidfuzz
rapidocr-onnxruntime
ratelimiter.post0
readme_renderer
realtime
referencing
regex
requests
requests-oauthlib
requests-toolbelt
retry
rfc3986
rich
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
ruff
SecretStorage
selenium
service-identity
setuptools
shapely
shellingham
six
smmap
sniffio
sortedcontainers
soupsieve
SpeechRecognition
SQLAlchemy
sqlparse
stack-data
starlette
storage3
streamlit
StrEnum
supabase
supafunc
sympy
tabulate
tenacity
termcolor
tiktoken
tokenizers
toml
tornado
tqdm
traitlets
trio
trio-websocket
tweepy
twine
Twisted
txaio
typer
types-requests
typing-inspect
typing_extensions
tzdata
unstructured
unstructured-client
uritemplate
urllib3
uvicorn
uvloop
vine
watchdog
watchfiles
wcwidth
websocket-client
websockets
wheel
whitenoise
wikipedia
wrapt
wsproto
XlsxWriter
yarl
zipp
zope.interface
